Added State Machine
Still a work in progress. This is meant to be more of a test on using a state machine to decide player actions as it could lead to more reusable code and also make it easier to add new abilities. Will add comments later to explain everything ... (If i don't, remind me)