Added Components
Testing out a component style of writing code. Could be better for scalability and reusing code. This is just a test for the workflow, I'm trying to experiment with different styles to get used to game development in godot. We'll discuss an overall method in the next discussion... Note: DO NOT MERGE TO MAIN YET!!!